🧩 :: RecruitmentDetailLabel 수정
